set authorizations in inventory service

+++ backend/inventory_mService/src/main/java/com/nahuelgg/inventory_app/inventories/controllers/InventoryController.java
@@ -6,6 +6,7 @@ import java.util.UUID;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 
 import com.nahuelgg.inventory_app.inventories.dtos.InventoryDTO;
@@ -13,16 +14,18 @@ import com.nahuelgg.inventory_app.inventories.dtos.ProductInInvDTO;
 import com.nahuelgg.inventory_app.inventories.dtos.ProductInputDTO;
 import com.nahuelgg.inventory_app.inventories.dtos.ProductToCopyDTO;
 import com.nahuelgg.inventory_app.inventories.dtos.UserFromUsersMSDTO;
+import com.nahuelgg.inventory_app.inventories.services.AuthorizationService;
 import com.nahuelgg.inventory_app.inventories.services.InventoryService;
 
+import lombok.RequiredArgsConstructor;
+
 @Controller
+@RequiredArgsConstructor
 public class InventoryController {
   private final InventoryService service;
+  private final AuthorizationService authorizationService;
 
-  public InventoryController(InventoryService service) {
-    this.service = service;
-  }
-
+  // Queries
   @QueryMapping
   public InventoryDTO getById(@Argument String id) {
     return service.getById(UUID.fromString(id));
@@ -45,13 +48,20 @@ public class InventoryController {
     return service.searchProductsInInventories(name, brand, model, categories, UUID.fromString(name));
   }
 
+  // Basic mutations
   @MutationMapping
   public InventoryDTO create(@Argument String name, @Argument String accountId) {
+    if (!authorizationService.checkUserIsAdmin())
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+
     return service.create(name, UUID.fromString(accountId));
   }
 
   @MutationMapping
   public String edit(@Argument String invId, @Argument String name) {
+    if (!authorizationService.checkUserIsAdmin())
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+
     if (service.edit(UUID.fromString(invId), name))
       return "Nombre editado con éxito";
     else return "No se pudo editar";
@@ -59,36 +69,59 @@ public class InventoryController {
 
   @MutationMapping
   public boolean delete(@Argument String id) {
+    if (!authorizationService.checkUserIsAdmin())
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+
     return service.delete(UUID.fromString(id));
   }
 
   @MutationMapping
   public boolean deleteByAccountId(@Argument String id) {
+    if (!authorizationService.checkUserIsAdmin())
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+
     return service.deleteByAccountId(UUID.fromString(id));
   }
 
+  // User mutations
   @MutationMapping
   public boolean addUser(@Argument UserFromUsersMSDTO user, @Argument String invId) {
+    if (!authorizationService.checkUserIsAdmin())
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+  
     return service.addUser(user, UUID.fromString(invId));
   }
 
   @MutationMapping
   public boolean removeUser(@Argument String userId, @Argument String accountId) {
+    if (!authorizationService.checkUserIsAdmin())
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+  
     return service.removeUser(UUID.fromString(userId), UUID.fromString(accountId));
   }
 
+  // Product mutations
   @MutationMapping
   public ProductInInvDTO addProduct(@Argument ProductInputDTO product, @Argument String invId) {
+    if (!authorizationService.checkUserHasPerm("addProducts", invId))
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+  
     return service.addProduct(product, UUID.fromString(invId));
   }
 
   @MutationMapping
   public boolean copyProducts(@Argument List<ProductToCopyDTO> products, @Argument String idTo) {
+    if (!authorizationService.checkUserHasPerm("addProducts", idTo))
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+  
     return service.copyProducts(products, UUID.fromString(idTo));
   }
 
   @MutationMapping
   public boolean editStockOfProduct(@Argument int relativeNewStock, @Argument String productRefId, @Argument String invId) {
+    if (!authorizationService.checkUserHasPerm("editInventory", invId))
+      throw new AccessDeniedException("No tiene permisos para realizar esta acción");
+
     return service.editStockOfProduct(relativeNewStock, UUID.fromString(productRefId), UUID.fromString(invId));
   }
 }
